feat(compute-js): extract crawlerHandlers module + add /@username OG handler
Refactors bech32 helpers and all crawler-OG handlers out of the 700-line
index.js into focused modules (bech32.js, crawlerHandlers.js) and adds
handleAtUsernameOg so social-media crawlers hitting /@username get proper
Open Graph / Twitter Card HTML with profile data from Funnelcake.
